windows build error fixed

+++ common/framework/MyRenderer.cpp
@@ -70,10 +70,10 @@ void MyRenderer::_writePixel(int x, int y, glm::vec4 color, float gama) {
     color.b = std::powf(color.b, gama);
   }
 
-  uint8_t r = 255.5f * std::clamp(color.r, 0.0f, 1.0f);
-  uint8_t g = 255.5f * std::clamp(color.g, 0.0f, 1.0f);
-  uint8_t b = 255.5f * std::clamp(color.b, 0.0f, 1.0f);
-  uint8_t a = 255.5f * std::clamp(color.a, 0.0f, 1.0f);
+  uint8_t r = uint8_t(255.5f * std::clamp(color.r, 0.0f, 1.0f));
+  uint8_t g = uint8_t(255.5f * std::clamp(color.g, 0.0f, 1.0f));
+  uint8_t b = uint8_t(255.5f * std::clamp(color.b, 0.0f, 1.0f));
+  uint8_t a = uint8_t(255.5f * std::clamp(color.a, 0.0f, 1.0f));
 
   uint32_t index = y * mFrameWidth + x;
   uint32_t colorValue = SDL_MapRGBA(mSurface->format, r, g, b, a);
